requirements.txt
whitenoise==3.2
airflow==1.8.0
alembic==0.8.10
apiai==1.2.3
appdirs==1.4.0
astroid==1.5.2
autopep8==1.3.2
backports.functools-lru-cache==1.4
beautifultable==0.3.0
bs4==0.0.1
dj-database-url==0.3.0
dj-static==0.0.6
Unipath==1.0
python-decouple==3
Pillow==3.3.0
Markdown==2.6.6
bleach==1.4.3
psycopg2==2.6.1
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.4
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
pytz==2017.2
requests==2.18.3
uritemplate==3.0.0
urllib3==1.22
facebook-sdk==2.0.0
terminaltables==3.1.0
Scrapy==0.24.4
